Added split displays and created files for the different screens
